Fixed issue with different screen resolutions
So we set it so that it is pixel perfect divisible on the width, then
set the height of each equal to that. There will almost always be
leftover, which is just placed outside of the maze on top and bottom.
It is less distracting than trying to spread the difference out. Only
other solution would be to implement an algorithm that would change the
height of different maze blocks, so they would all be different heights
to even out the difference.